Funcionando correctamente

+++ controllers/controllersAdmin/adminController.js
@@ -6,14 +6,26 @@ const ModelItebis = require("../../models/modelAdmin/itebis");
 const ModelProducto = require("../../models/modelComercios/producto");
 const ModelPedido = require("../../models/modelCliente/pedido");
 const verific = require("../../utils/verificUserLog");
+const bcrypt = require("bcryptjs");
 
 exports.getClientes = async (req, res, next) => {
+  const role = req.session.user.role;
+  console.log("El rol del admin: " + role);
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   verific(req, res, next);
   //   el nombre, el apellido,
   // la cantidad de pedidos que ha realizado ese cliente, el teléfono y el correo.
 
   const items = await ModelCliente.findAll();
-  const clientes = await items.map((cliente) => cliente.dataValues);
+  const clientes = items.map((cliente) => {
+    data = cliente.dataValues;
+    data.isActivo = cliente.active;
+    return data;
+  });
+
   res.render("viewsAdmin/listadoCliente", {
     pageTitle: "Food Rush | Admin",
     layout: "layoutAdmin",
@@ -22,6 +34,12 @@ exports.getClientes = async (req, res, next) => {
   });
 };
 exports.getViewItebis = async (req, res, next) => {
+  const role = req.session.user.role;
+  console.log("El rol del admin: " + role);
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   // const idAdmin = verific(req, res, next);
   const itbis = await ModelItebis.findOne();
   console.log(" El itbis ", itbis);
@@ -33,7 +51,82 @@ exports.getViewItebis = async (req, res, next) => {
     editMode: false,
   });
 };
+
+exports.adminEditar = async (req, res, next) => {
+  const role = req.session.user.role;
+  console.log("El rol del admin: " + role);
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
+  const adminId = req.params.adminId;
+
+  const admin = await ModelAdministrador.findOne({ where: { id: adminId } });
+
+  res.render("viewsAdmin/viewEditarAdmin", {
+    pageTitle: "Food Rush | Ediat",
+    layout: "layoutAdmin",
+    Admin: admin.dataValues,
+  });
+};
+exports.editarAdminPost = async (req, res, next) => {
+  const role = req.session.user.role;
+  console.log("El rol del admin: " + role);
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
+  const {
+    name,
+    lastName,
+    identification,
+    email,
+    user,
+    password,
+    confirmPassword,
+    adminId,
+  } = req.body;
+
+  if (password !== confirmPassword) {
+    req.flash("errors", "Passwords do not match");
+    console.log("Passwords do not match");
+    return res.redirect("back");
+  }
+
+  try {
+    const existingUser = await ModelAdministrador.findOne({
+      where: { user: user },
+    });
+
+    const hashedPassword = await bcrypt.hash(password, 12);
+
+    await ModelAdministrador.update(
+      {
+        name,
+        lastName,
+        identification,
+        email,
+        user,
+        password: hashedPassword,
+      },
+      { where: { id: adminId } }
+    );
+
+    console.log("Registro correcto");
+    return res.redirect("/Admin/mantenimientoAdmin");
+  } catch (err) {
+    console.log("Este es el error ", err);
+    req.flash("errors", "Something went wrong, please try again later");
+    return res.redirect("back");
+  }
+};
+
 exports.postEdidtViewItebis = async (req, res, next) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   const idAdmin = verific(req, res, next);
   const itbis = req.body.itbis;
   await ModelItebis.update(
@@ -44,18 +137,71 @@ exports.postEdidtViewItebis = async (req, res, next) => {
   );
   return res.redirect("back");
 };
+exports.postActDes = async (req, res, next) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
+  const id = req.body.id;
+  const rol = req.body.rol;
+
+  switch (rol) {
+    case "cliente":
+      const clienteitem = await ModelCliente.findOne({ where: { id: id } });
+      if (clienteitem.dataValues.active == true) {
+        await ModelCliente.update({ active: false }, { where: { id: id } });
+      } else await ModelCliente.update({ active: true }, { where: { id: id } });
+
+      break;
+    case "delivery":
+      const deliveryitem = await ModelDelivery.findOne({ where: { id: id } });
+      if (deliveryitem.dataValues.active == true) {
+        await ModelDelivery.update({ active: false }, { where: { id: id } });
+      } else
+        await ModelDelivery.update({ active: true }, { where: { id: id } });
+      break;
+    case "administrador":
+      const adminitem = await ModelAdministrador.findOne({ where: { id: id } });
+      if (adminitem.dataValues.active == true) {
+        await ModelAdministrador.update(
+          { active: false },
+          { where: { id: id } }
+        );
+      } else
+        await ModelAdministrador.update(
+          { active: true },
+          { where: { id: id } }
+        );
+      break;
+    case "comercio":
+      const comercioitem = await ModelComercio.findOne({ where: { id: id } });
+      if (comercioitem.dataValues.active == true) {
+        await ModelComercio.update({ active: false }, { where: { id: id } });
+      } else
+        await ModelComercio.update({ active: true }, { where: { id: id } });
+      break;
+  }
+
+  return res.redirect("back");
+};
 
 exports.getDeliveries = async (req, res) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   try {
     //     el nombre, el apellido,
     // la cantidad de pedidos que ha entregado ese delivery, el teléfono y el correo.
     const items = await ModelDelivery.findAll();
     console.log("El items: ", items);
-    const deliveries = items.map((deli) => deli.dataValues);
-    deliveries.forEach((element) => {
-      console.log("El delivery: ", element);
+    const deliveries = items.map((deli) => {
+      data = deli.dataValues;
+      data.isActivo = deli.active;
+      return data;
     });
-
     res.render("viewsAdmin/listadoDelivery", {
       pageTitle: "Food Rush | Admin",
       layout: "layoutAdmin",
@@ -68,6 +214,11 @@ exports.getDeliveries = async (req, res) => {
 };
 
 exports.getComercios = async (req, res, next) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   //   el nombre del
   // comercio, el logo del comercio, la cantidad de pedidos de ese comercio que hay, el teléfono,
   // la hora de apertura, la hora de cierre y el correo.
@@ -88,10 +239,22 @@ exports.getComercios = async (req, res, next) => {
   });
 };
 
-exports.getAdministradores = async (req, res) => {
+exports.getAdministradores = async (req, res, next) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   try {
+    const idAdmin = verific(req, res, next);
     const items = await ModelAdministrador.findAll();
-    const admis = items.map((admin) => admin.dataValues);
+    const admis = items.map((admin) => {
+      data = admin.dataValues;
+      data.isActivo = admin.active;
+      data.im = admin.id === idAdmin;
+      console.log("Es el mismo? ", data.im);
+      return data;
+    });
 
     res.render("viewsAdmin/listadoAdmins", {
       pageTitle: "Food Rush | Admin",
@@ -119,6 +282,11 @@ exports.getAdministradores = async (req, res) => {
 // };
 
 exports.createAdministrador = async (req, res) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   try {
     const { name, lastName, identification, email, user, password } = req.body;
     await Administrador.create({
@@ -136,6 +304,11 @@ exports.createAdministrador = async (req, res) => {
 };
 
 exports.updateAdministrador = async (req, res) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   try {
     const { id } = req.params;
     const { name, lastName, identification, email, user, password } = req.body;
@@ -150,6 +323,11 @@ exports.updateAdministrador = async (req, res) => {
 };
 
 exports.deleteAdministrador = async (req, res) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   try {
     const { id } = req.params;
     await Administrador.destroy({ where: { id } });
@@ -159,6 +337,11 @@ exports.deleteAdministrador = async (req, res) => {
   }
 };
 exports.logout = (req, res) => {
+  const role = req.session.user.role;
+  if (role != "administrador") {
+    req.flash("errors", "Usted no tiene acceso a esta area, buen loco.");
+    return res.redirect("/login");
+  }
   req.session.destroy();
   res.redirect("/login");
 };
